Fix: typescript is a lazy mother trucker. (v.0.1.1)
When we use PropsObj, typescript assumes that we are communicating
Important Information via this name when actually we're just trying to
save a few bytes. So downstream everyone sees something messy like

    Typeable<PropsObj<{
      files: Typeable<Record<string, PropsObj<{
        hash: Typeable<string>
      }>>>
    }>>

rather than the much more readable and obvious

    Typeable<{files: Record<string, {hash: string}>}>

and that's one of the simplest examples!

If we inline it, then TypeScript magically becomes less lazy.